+++ src/main/java/umc/animore/model/Location.java
@@ -18,10 +18,10 @@ public class Location {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "location_id")
-    private Integer locationId;
+    private Long locationId;
     @Column(name = "latitude")
-    private double latitude;
+    private Double latitude;
     @Column(name = "longitude")
-    private double longitude;
+    private Double longitude;
 
 }
