+++ src/main/java/umc/animore/repository/SearchHistoryRepository.java
@@ -6,8 +6,8 @@ import umc.animore.model.SearchHistory;
 
 import java.util.List;
 
-public interface SearchHistoryRepository extends JpaRepository<SearchHistory, Integer> {
+public interface SearchHistoryRepository extends JpaRepository<SearchHistory, Long> {
 
     //최근 검색기록 - 3개
-    List<SearchHistory> findByUserIdxOrderBySearchCreateAtDesc(int userIdx);
+    List<SearchHistory> findByUserIdxOrderBySearchCreateAtDesc(Long userIdx);
 }
