+++ src/main/java/umc/animore/repository/TownRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import umc.animore.model.Town;
 
 @Repository
-public interface TownRepository extends JpaRepository<Town, Integer> {
+public interface TownRepository extends JpaRepository<Town, Long> {
 
     Town getTownIdByCityAndDistrict(String city, String district); //town_id찾기
 }
